+++ Sources/Entities/MessageImpl.swift
@@ -12,9 +12,26 @@ import Foundation
 import PubNubChat
 import PubNubSDK
 
+/// A concrete implementation of the ``Message`` protocol.
+///
+/// This class should be used as a ready-to-use solution for most use cases requiring the features defined by the ``Message`` protocol. Also, this class inherits all the documentation
+/// for methods defined in the ``Message`` protocol. Refer to the ``Message`` protocol for details on how individual methods work.
 public final class MessageImpl {
   let target: BaseMessage<PubNubChat.Message>
 
+  /// Creates a new ``MessageImpl`` instance
+  ///
+  /// - Parameters:
+  ///   - chat: Reference to the main Chat object
+  ///   - timetoken: Timetoken for the message
+  ///   - content: Original text content of the message
+  ///   - channelId: Unique identifier for the channel in which the message was sent
+  ///   - userId: Unique ID of the user who sent the message
+  ///   - actions: Any actions associated with the message, such as reactions, replies, or other interactive elements
+  ///   - meta: Extra information added to the message giving additional context
+  ///   - mentionedUsers: List of mentioned users with IDs and names
+  ///   - referencedChannels: List of referenced channels with IDs and names
+  ///   - quotedMessage: Access the original quoted message
   public convenience init(
     chat: ChatImpl,
     timetoken: Timetoken,
