+++ Sources/Entities/ThreadChannelImpl.swift
@@ -12,9 +12,25 @@ import Foundation
 import PubNubChat
 import PubNubSDK
 
+/// A concrete implementation of the ``ThreadChannelImpl`` protocol.
+///
+/// This class should be used as a ready-to-use solution for most use cases requiring the features defined by the ``ThreadChannelImpl`` protocol. Also, this class inherits all the documentation
+/// for methods defined in the ``ThreadChannelImpl`` protocol. Refer to the ``ThreadChannelImpl`` protocol for details on how individual methods work.
 public final class ThreadChannelImpl {
   let target: BaseChannel<PubNubChat.ThreadChannel, PubNubChat.ThreadMessage>
 
+  /// Creates a new ``ThreadChannelImpl`` object
+  ///
+  /// - Parameters:
+  ///   - parentMessage: Message for which the thread was created
+  ///   - chat: Reference to the main Chat object
+  ///   - id: Unique identifier for the channel
+  ///   - name: Display name or title of the channel
+  ///   - custom: Any custom properties or metadata associated with the channel in the form of a map of key-value pairs
+  ///   - description: Brief description or summary of the channel's purpose or content
+  ///   - updated: The last updated timestamp for the object
+  ///   - status: Current status of the channel, like online, offline, or archived
+  ///   - type: Represents the type of channel
   public convenience init(
     parentMessage: MessageImpl,
     chat: ChatImpl,
@@ -24,8 +40,7 @@ public final class ThreadChannelImpl {
     description: String? = nil,
     updated: String? = nil,
     status: String? = nil,
-    type: ChannelType? = nil,
-    threadCreated: Bool = true
+    type: ChannelType? = nil
   ) {
     let underlyingThreadChannel = PubNubChat.ThreadChannelImpl(
       parentMessage: parentMessage.target.message,
@@ -38,7 +53,7 @@ public final class ThreadChannelImpl {
       updated: updated,
       status: status,
       type: type?.transform(),
-      threadCreated: threadCreated
+      threadCreated: true
     )
     self.init(
       channel: underlyingThreadChannel
